getCirros uses case switch for variable launchSetup now borrows getCirros and createFlavors scripts instead of rewriting them

+++ scripts/overcloud/getCirros.sh
@@ -1,20 +1,30 @@
-# +------+
-# | curl |
-# +------+
+# Valid formats are RAW or QCOW2
+FORMAT=QCOW2
 
-curl -O http://download.cirros-cloud.net/0.3.5/cirros-0.3.5-x86_64-disk.img
+source /home/stack/overcloudrc
 
-# +------+
-# | wget |
-# +------+
+if [ ! -d /home/stack/images ]; then mkdir -p /home/stack/images; fi
+cd /home/stack/images
 
-# wget http://download.cirros-cloud.net/0.3.5/cirros-0.3.5-x86_64-disk.img
+if [ ! -f cirros-0.3.5-x86_64-disk.img ]; then
+  curl -O http://download.cirros-cloud.net/0.3.5/cirros-0.3.5-x86_64-disk.img
+fi
 
-# +-------------------------------------+
-# | Convert image if using Ceph storage |
-# +-------------------------------------+
-
-# qemu-img convert -f qcow2 -O raw cirros-0.3.5-x86_64-disk.img cirros-0.3.5-x86_64-disk.raw
-# openstack image create --disk-format raw --container-format bare --file cirros-0.3.5-x86_64-disk.raw cirros
-
-openstack image create --disk-format qcow2 --container-format bare --file cirros-0.3.5-x86_64-disk.img cirros
+case $FORMAT in
+  RAW)
+    # +-------------------------------------+
+    # | Convert image if using Ceph storage |
+    # +-------------------------------------+
+    if [ ! -f cirros-0.3.5-x86_64-disk.raw ]; then
+      qemu-img convert -f qcow2 -O raw cirros-0.3.5-x86_64-disk.img cirros-0.3.5-x86_64-disk.raw
+    fi
+    openstack image create --disk-format raw --container-format bare --file cirros-0.3.5-x86_64-disk.raw cirros
+  ;;
+  QCOW2)
+    openstack image create --disk-format qcow2 --container-format bare --file cirros-0.3.5-x86_64-disk.img cirros
+  ;;
+  *)
+    echo "INVALID FORMAT."
+    exit 1
+  ;;
+esac
